requirements-dev.txt
numpy
pytest
matplotlib
cupy-cuda12x
